Generalize CUDA detection in settings.{sh,csh}
CUDA toolkit is not supplied as a Thirdparty package, so remove all mentions
of CUDA_SYSTEM variable.

If CUDA compiler (nvcc) is available in PATH, use it to obtain the location of
CUDA toolkit, but do not override any user settings.  Do not hardcode any
possible path in /usr/local as a guess of the location.

Make settings.sh and settings.csh follow each other more closely.

Use lib64 instead of lib as the default expected library subdirectory.