Speed up file modification checks in parallel runs
Avoid many simulaneous requests to the (usually shared, networked) filesystem
when checking modification times of files that can be re-read at runtime
(enabled with "runTimeModifiable true" in controlDict).  Instead, first check
the modification only from the master process, and proceed to check
modification time on other ranks only if the master has noticed the update.

This cannot cause fewer runtime file updates, as in both functions changed
with the patch the next action is to verify that all ranks have newer
modification times, so if the master did not see the update, files would not
be re-read anyway.  However, warnings when one of slave nodes noticed an
update before the master did will not be issued anymore.